[hardware] Move Flash and friends into their own module
