+++ src/hardware/flash.rs
@@ -2,47 +2,21 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-//! Pluggable hardware functionality
+//! External, remote flash abstraction.
 //!
-//! This module provides traits for plugging in OS calls to specialized
-//! hardware functions, such as power-on and reset-related capabilities.
-//! `manticore` uses this functionality to respond to certain protocol
-//! requests.
-
-use core::time::Duration;
+//! This crate provides the [`Flash`] (and related) traits, which represent
+//! *abstract flash devices*. An abstract flash device is a region of memory
+//! that can be transactionally read or written. Such a "device" can range
+//! from a simple Rust slice to a remote SPI flash device (or even a subregion of
+//! it!).
+//!
+//! [`Flash`]: trait.Flash.html
 
 use zerocopy::AsBytes;
 use zerocopy::FromBytes;
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
-
-/// Provides access to "chip identity" information of various types.
-pub trait Identity {
-    /// Returns a string indicating the RoT's firmware version.
-    ///
-    /// Although not enforced, it is recommended that this be an ASCII string.
-    fn firmware_version(&self) -> &[u8; 32];
-
-    /// Returns the "unique device identity" for the device. This is a binary
-    /// value of unspecified format.
-    fn unique_device_identity(&self) -> &[u8];
-}
-
-/// Provides access to device reset-related information for a particular
-/// device.
-pub trait Reset {
-    /// Returns the number of times the device has been reset since it was
-    /// powered on.
-    fn resets_since_power_on(&self) -> u32;
-
-    /// Returns the uptime of the device, i.e., the absolute duration since it
-    /// was last released from reset.
-    ///
-    /// The resolution and accuracy of this value are expected to be
-    /// best-effort.
-    fn uptime(&self) -> Duration;
-}
 
 /// Provides access to a flash-like storage device.
 ///
@@ -51,8 +25,7 @@ pub trait Reset {
 /// to implement these operations efficiently with respect to the underlying
 /// device.
 ///
-/// `Flash` is also implemented on main-memory `u8` buffers, which can be used
-/// for testing.
+/// The `Flash` trait comes implemented for `[u8]`.
 pub trait Flash {
     /// The error type returned by transactions with this `Flash`.
     type Error: Sized;
@@ -87,7 +60,7 @@ pub struct FlashPtr {
 
 impl FlashPtr {
     /// Convenience method for creating a `FlashPtr` without having to use
-    /// a struct literal
+    /// a struct literal.
     pub const fn new(address: u32) -> Self {
         Self { address }
     }
@@ -145,65 +118,5 @@ impl Flash for [u8] {
         }
         out.copy_from_slice(&self[start..end]);
         Ok(())
-    }
-}
-
-#[cfg(test)]
-pub(crate) mod fake {
-    use core::convert::TryInto;
-    use core::time::Duration;
-
-    /// A fake `Identity` that returns fixed values.
-    pub struct Identity {
-        firmware_version: Vec<u8>,
-        unique_id: Vec<u8>,
-    }
-
-    impl Identity {
-        /// Creates a new `fake::Identity`.
-        pub fn new(firmware_version: &[u8], unique_id: &[u8]) -> Self {
-            let mut firmware_version = firmware_version.to_vec();
-            while firmware_version.len() < 32 {
-                firmware_version.push(0);
-            }
-            firmware_version.truncate(32);
-
-            Self {
-                firmware_version,
-                unique_id: unique_id.to_vec(),
-            }
-        }
-    }
-
-    impl super::Identity for Identity {
-        fn firmware_version(&self) -> &[u8; 32] {
-            self.firmware_version[..32].try_into().unwrap()
-        }
-        fn unique_device_identity(&self) -> &[u8] {
-            &self.unique_id[..]
-        }
-    }
-
-    /// A fake `Reset` that returns fixed values.
-    pub struct Reset {
-        resets: u32,
-        uptime: Duration,
-    }
-
-    impl Reset {
-        /// Creates a new `fake::Reset`.
-        pub fn new(resets: u32, uptime: Duration) -> Self {
-            Self { resets, uptime }
-        }
-    }
-
-    impl super::Reset for Reset {
-        fn resets_since_power_on(&self) -> u32 {
-            self.resets
-        }
-
-        fn uptime(&self) -> Duration {
-            self.uptime
-        }
     }
 }
